[REVIEW] R53772121: Don't auto scroll when editing (#95)
* Prevent autoscrolling for form items that have already been answered

This PR adds a new ivar to `ORKFormStepViewController.m` that is used to track which sections have been answered already. When a question has already been answered, autoscrolling will be disabled. This change is necessary to make editing previously answered questions a good experience.

Alternatives Considered:
I tried to find an implementation that did not require adding any new state variables, but the nature of didSelectRowAtIndex made that too difficult. At the point where the decision to autoscroll is made (in didSelectCellAtIndexPath), we need to know if the cell was previously selected or not. We cannot do this simply by looking at the current state of the cells because in `didSelectRow` the cell has already become selected and does not retain memory of its previous state, so I wound up falling back on a stateful solution.

* Support for persisting state

* Fixed issue in state restoration

* Removed now redundant code

* Cleaned up formatting