Fix crash when using a UIScene based app
Crash resulted from trying to access the window property on AppDelegate. In UIScene based app, the window does not reside on the delegate, but rather on the new scene delegate. This PR provides a default value if there is no app delegate.

I also had to extract the setting of colors out of the setup methods. Because the setup methods run before the view is added to a window, the tint colors will not be set correctly. Deferring setting of the colors until the window is available should fix this problem.

This PR will require thorough QA to make sure nothing has broken, especially in ResearchApp.