rdar://65995162 (Terminating app due to uncaught exception 'NSInvalidUnarchiveOperationException', reason: 'The data couldn’t be read because it isn’t in the correct format.') (#585)
Trying to use NSSecureCoding to decode archived ORKQuestionResults with ORKDontKnowAnswer crashes. This is because the decoding implementations are, for example, like this:

- (instancetype)initWithCoder:(NSCoder *)aDecoder {
    self = [super initWithCoder:aDecoder];
    if (self) {
        ORK_DECODE_OBJ_CLASS(aDecoder, textAnswer, NSString);
    }
    return self;
}

so secure decoding doesn't know the object can be of ORKDontKnowAnswer type.

This adds a new `ORKQuestionType.noAnswerType` property of `ORKNoAnswer` type, and then makes `ORKDontKnowAnswer` a subclass of `ORKNoAnswer` (in case we want to differentiate between other special cases such as "Don't know" vs "Prefer not to answer" or "Skipped" in the future). Internal to ResearchKit, both a normal answer or a special case answer are stored in the same `_typedAnswerOrNoAnswer` internal variable, and both are served through the private `-answer` method, but the public interface has changed so both the `resultSpecificAnswer` and the `noAnswerType` properties are type-safe.

Secure decoding maintains a fall-back to the old property keys in the subclasses, to maintain backwards compatibility when decoding already archived objects.

This also changes the serialized result from:

{"_class":"ORKTextQuestionResult","endDate":"2019-05-27T00:35:06-0700","startDate":"2019-05-27T00:35:06-0700","identifier":"","questionType":0,"userInfo":{},"noAnswer_textAnswer":{"_class":"ORKDontKnowAnswer"}}

to

{"_class":"ORKChoiceQuestionResult","endDate":"2019-05-27T00:35:06-0700","startDate":"2019-05-27T00:35:06-0700","identifier":"","questionType":0,"userInfo":{},"noAnswerType":{"_class":"ORKDontKnowAnswer"}}
for all result subclasses, and saves a lot of JSON deserialization special cases.