Ensure that ORKVideoInstructionStepResult is JSON compatible
Currently, if the user does not watch the video, the step result that is produced has a playbackStoppedTime with a value of NaN. JSON doesn't have an affordance for expressing NaN, so this leads to ORKESerialization crashing at run time.

The solution proposed here is to use FLT_MAX to represent unplayed videos instead of NAN. Values of -1 and 0 were considered as well, but 0 already has another interpretation and consumers may not expect negative numbers like -1. We chose FLT_MAX because it is already used to express a similar concept in some of the hearing tests, making it consistent with other use cases in ResearchKit.