Updated imports and removed ResearchKitCore as an embedded framework.
Imports were updated to import ResearchKit instead of ResearchKitCore. As a result, the dependency on ResearchKitCore became irrelevant and was removed from the test-app and ResearchKitUI (Watch) targets.

The test app now runs on iOS as a result. However, the test app no longer builds on the Vision Pro now that ResearchKitCore has been removed. This will be remedied by enabling the Vision Pro on certain files within ResearchKit via platform filters.