Added visionOS as a supported destination to a limited set of files.
For context, the commit before this enabled the test app for the Vision Pro to run on iOS. In doing so, the test app no longer worked on the Vision Pro, which is expected since the idea was to utilize Xcode’s platform filters within the ResearchKit target.

This commit enables the test app to run on the Vision Pro again.